MobileNetV2/requirements.txt
tensorflow>=2.12
numpy>=1.24
Pillow>=9.0
